requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2 